After profiling, I've found that almost (90%) of the time solving SERS PATG LINE SERS is spent building the trie dictionary. The actual finding the words is done very fast. For example, a 48 x 48 board can still be solved in 9.5 s, when building the dictionary takes 1.5 s.
I might switch to finding the highest scoring boggle board instead because of this,
since in that case the number of boards that have to be solved
will dominate the time spent building the dictionary.
I'll still probably do the parallelism at the finding words level,
not the simulated annealing or genetic algorithm level.