Got it to work finally!  There's a bug in mkStdGen.
mkStdGen takes an Int seed, a 64-bit int, but internally,
it truncates it to a 32-bit int,
but it doesn't seem to use 2s complement truncation,
so large 64-bit ints kept turning into a 32-bit seed of 1,
which meant the StdGen was the same every time.

Now I'm purposefully restricting the seed range to Int32,
so the simulated annealing and randomness all work now.

I still have to tweak the simulated annealing params though,
since it's taking a bit too long and it already optimizes quicker.