Using the remaining subDict size as a proxy for load balancing, i.e. only parallelizing when the remaining dict size is large enough because then there's probably enough work for parallelism to help.
This makes the overall program only slightly slower than before,
but it creates a speedup of ~2x instead of just over 1x before.
This is because Trie.size is O(n), so in the serial version,
it's not worth it, meaning the serial version is much slower.
But in the parallel version, Trie.size and parallelism cancel each other out.

Thus, if I can optimize, cache, or pre-compute the trie size,
then I should get a significant speedup.