Pre-computed Trie.size by patching bytestring-trie (I'm using my own fork of it now) to export the data constructors, which means I can write more specialized algorithms in my code. Trie.ofSizes stores the sub-trie's size in the trie, so finding the size of a trie is now O(1). It may have to iterate over the top-level branches, but there are a constant bound on those.
This sped up the serial version back to normal,
and helped the parallel version somewhat,
not nearly as much as I'd hoped.

Doing it this way with the remaining dict size to determine
if more parallelism should be sparked, however,
does lead to much less wasted sparks.
Now about 2/3 of the sparks are converted,
compared to very little before.